requirements.txt
langchain
langchain-openai
openai

google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

google-auth-oauthlib 
google-auth-httplib2  
google-api-python-client

googlemaps

requests

python-dotenv

dateparser